pyproject.toml
[tool.poetry]
name = "assume_py"
version = "0.1.0"
description = "An AWS CLI session switch user(su) command."
authors = ["Gatewaynode"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.23.2"
botocore = "^1.26.2"
cffi = "^1.15.0"
click = "^8.1.3"
cryptography = "^37.0.2"
Jinja2 = "^3.1.2"
jmespath = "^1.0.0"
MarkupSafe = "^2.1.1"
mypy-extensions = "^0.4.3"
pathspec = "^0.9.0"
platformdirs = "^2.5.2"
pycparser = "^2.21"
Pygments = "^2.12.0"
python-dateutil = "^2.8.2"
PyYAML = "^6.0"
s3transfer = "^0.5.2"
six = "^1.16.0"
tomli = "^2.0.1"
urllib3 = "^1.26.9"
xdg = "^5.1.1"


[tool.poetry.dev-dependencies]
black = "^22.3.0"
pdoc = "^12.0.1"

[tool.poetry.scripts]
as = 'assume_py.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
